rename sales_manager->salesManager
