+++ src/salesManager.js
